+++ app/Http/Controllers/TweetController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Tweet;
+use App\Models\Comments;
 
 class TweetController extends Controller
 {
@@ -13,19 +14,31 @@ class TweetController extends Controller
         return response()->json($tweets);
     }
 
+    public function show($id)
+    {
+        $tweet = Tweet::find($id);
+        $coments = Comments::where('tweets_id', $id)->get();
+        return response()->json([
+            'tweet' => $tweet,
+            'comments' => $coments
+            ]);
+
+    }
+
     public function create()
     {
         try {
             $rules = [
-                'user_id' => 'required',
                 'description' => 'required',
             ];
             $messages = [
-                'user_id.required' => 'User ID is required',
                 'description.required' => 'Tweet is required',
             ];
             $this->validate(request(), $rules, $messages);
-            $tweet =Tweet::create(request()->all());
+            $tweet =Tweet::create([
+                'user_id' => auth()->user()->id,
+                'description' => request()->description,
+            ]);
             return response()->json($tweet);
         } catch (\Exception $e) {
             return response()->json($e->getMessage());
@@ -33,13 +46,18 @@ class TweetController extends Controller
     }
 
     public function update($id){
+
+        $tweet = Tweet::find($id);
+
+        if($tweet->user_id != auth()->user()->id){
+            return response()->json('You are not authorized to update this tweet');
+        }
+        
         try {
             $rules = [
-                'user_id' => 'required',
                 'description' => 'required',
             ];
             $messages = [
-                'user_id.required' => 'User ID is required',
                 'description.required' => 'Tweet is required',
             ];
             $this->validate(request(), $rules, $messages);
@@ -49,9 +67,15 @@ class TweetController extends Controller
         } catch (\Exception $e) {
             return response()->json($e->getMessage());
         }
+        
     }
 
     public function delete($id){
+        $tweet = Tweet::find($id);
+
+        if($tweet->user_id != auth()->user()->id){
+            return response()->json('You are not authorized to update this tweet');
+        }
         try {
             $tweet = Tweet::find($id);
             $tweet->delete();
@@ -60,4 +84,5 @@ class TweetController extends Controller
             return response()->json($e->getMessage());
         }
     }
+    
 }
